+++ packages/gamut/Markdown/overrides/Code/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Code = props => <code {...props} />;
+
+export default Code;