+++ packages/webpack-config/CHANGELOG.md
@@ -3,6 +3,90 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.10.1"></a>
+## [3.10.1](http://github.com/RyzacInc/gamut/packages/webpack-config/compare/@codecademy/webpack-config@3.10.0...@codecademy/webpack-config@3.10.1) (2019-01-14)
+
+
+### Bug Fixes
+
+* add more dev server headers to avoid local errors ([#156](http://github.com/RyzacInc/gamut/packages/webpack-config/issues/156)) ([51ff9fd](http://github.com/RyzacInc/gamut/packages/webpack-config/commit/51ff9fd))
+
+
+
+
+
+<a name="3.10.0"></a>
+# [3.10.0](http://github.com/RyzacInc/gamut/packages/webpack-config/compare/@codecademy/webpack-config@3.9.0...@codecademy/webpack-config@3.10.0) (2018-10-29)
+
+
+### Features
+
+* More dev server options ([#151](http://github.com/RyzacInc/gamut/packages/webpack-config/issues/151)) ([d181f5a](http://github.com/RyzacInc/gamut/packages/webpack-config/commit/d181f5a))
+
+
+
+
+
+<a name="3.9.0"></a>
+# [3.9.0](http://github.com/RyzacInc/gamut/packages/webpack-config/compare/@codecademy/webpack-config@3.8.0...@codecademy/webpack-config@3.9.0) (2018-10-25)
+
+
+### Features
+
+* switch back to dev server ([#149](http://github.com/RyzacInc/gamut/packages/webpack-config/issues/149)) ([335377e](http://github.com/RyzacInc/gamut/packages/webpack-config/commit/335377e))
+
+
+
+
+
+<a name="3.8.0"></a>
+# [3.8.0](http://github.com/RyzacInc/gamut/packages/webpack-config/compare/@codecademy/webpack-config@3.7.0...@codecademy/webpack-config@3.8.0) (2018-10-03)
+
+
+### Features
+
+* Better babel 7 support ([#145](http://github.com/RyzacInc/gamut/packages/webpack-config/issues/145)) ([49a1e23](http://github.com/RyzacInc/gamut/packages/webpack-config/commit/49a1e23))
+
+
+
+
+
+<a name="3.7.0"></a>
+# [3.7.0](http://github.com/RyzacInc/gamut/packages/webpack-config/compare/@codecademy/webpack-config@3.6.0...@codecademy/webpack-config@3.7.0) (2018-08-30)
+
+
+### Features
+
+* remove hot-loader as a dependency ([#146](http://github.com/RyzacInc/gamut/packages/webpack-config/issues/146)) ([b3b29b0](http://github.com/RyzacInc/gamut/packages/webpack-config/commit/b3b29b0))
+
+
+
+
+
+<a name="3.6.0"></a>
+# [3.6.0](http://github.com/RyzacInc/gamut/packages/webpack-config/compare/@codecademy/webpack-config@3.5.0...@codecademy/webpack-config@3.6.0) (2018-08-28)
+
+
+### Features
+
+* switch to webpack-serve ([#144](http://github.com/RyzacInc/gamut/packages/webpack-config/issues/144)) ([d5389b0](http://github.com/RyzacInc/gamut/packages/webpack-config/commit/d5389b0))
+
+
+
+
+
+<a name="3.5.0"></a>
+# [3.5.0](http://github.com/RyzacInc/gamut/packages/webpack-config/compare/@codecademy/webpack-config@3.4.0...@codecademy/webpack-config@3.5.0) (2018-08-21)
+
+
+### Features
+
+* updates for babel 7 ([748b911](http://github.com/RyzacInc/gamut/packages/webpack-config/commit/748b911))
+
+
+
+
+
 <a name="3.4.0"></a>
 # [3.4.0](http://github.com/RyzacInc/gamut/packages/webpack-config/compare/@codecademy/webpack-config@3.3.0...@codecademy/webpack-config@3.4.0) (2018-06-20)